+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.test.jsx
@@ -86,6 +86,54 @@ describe('src/components/custom-field-input-single-choice/custom-field-input-sin
     });
   });
 
+  describe('filtering', () => {
+    const choices = [{
+      id: '1',
+      label: 'Hey',
+    }, {
+      id: '2',
+      label: 'Hi',
+    }];
+
+    it('occurs when provided text to filter on', () => {
+      render(<CustomFieldInputSingleChoice {...requiredProps} choices={choices} />);
+      userEvent.click(screen.getByLabelText('Test label'));
+      userEvent.type(document.activeElement, 'he');
+
+      expect(screen.getByText('Hey')).toBeInTheDocument();
+      expect(screen.queryByText('Hi')).not.toBeInTheDocument();
+    });
+
+    it('allows filtering when focused on the component', () => {
+      render(<CustomFieldInputSingleChoice {...requiredProps} choices={choices} />);
+      userEvent.tab();
+      userEvent.type(document.activeElement, 'he', { skipClick: true });
+
+      expect(document.activeElement).toHaveValue('he');
+      expect(screen.queryByText('Hey')).toBeInTheDocument();
+      expect(screen.queryByText('Hi')).not.toBeInTheDocument();
+    });
+
+    it('respects deleting the value', () => {
+      render(<CustomFieldInputSingleChoice {...requiredProps} choices={choices} />);
+      userEvent.click(screen.getByLabelText('Test label'));
+      userEvent.click(screen.getByText('Hey'));
+      userEvent.click(screen.getByLabelText('Test label'));
+      fireEvent.change(document.activeElement, { target: { value: '' } });
+
+      expect(document.activeElement).toHaveValue('');
+    });
+
+    it('allows selection on filtering', () => {
+      render(<CustomFieldInputSingleChoice {...requiredProps} choices={choices} />);
+      userEvent.click(screen.getByLabelText('Test label'));
+      userEvent.type(document.activeElement, 'he');
+      userEvent.click(screen.getByText('Hey'));
+
+      expect(document.activeElement).toHaveValue('Hey');
+    });
+  });
+
   describe('id API', () => {
     it('accepts an ID', () => {
       render(<CustomFieldInputSingleChoice {...requiredProps} id="this-is-an-id" />);